Feat: implement WelcomeSlideUI inside WelcomeScreen
